+++ CHANGES.md
@@ -8,6 +8,7 @@ Fixed:
 
 - Disabled the startupapicheck job in cert-manager startup.
 - Fixed an install bug related to incorrect ingress names in certain configurations.
+- Fixed an ArgoCD bug to reuse Rancher auth tokens when applicable.
 
 Features: